Timeout removed from client
